+++ test/match_test.rb
@@ -59,8 +59,8 @@ class MatchTest < Minitest::Test
   end
 
   def test_stemming_tokens
-    assert_equal ["milk"], Product.search_index.tokens("milks", analyzer: "searchkick_search")
-    assert_equal ["milk"], Product.search_index.tokens("milks", analyzer: "searchkick_search2")
+    assert_equal ["milk"], Product.searchkick_index.tokens("milks", analyzer: "searchkick_search")
+    assert_equal ["milk"], Product.searchkick_index.tokens("milks", analyzer: "searchkick_search2")
   end
 
   # fuzzy
@@ -263,6 +263,8 @@ class MatchTest < Minitest::Test
     assert_search "fresh honey", ["fresh", "honey"], {operator: "or"}
     assert_search "fresh honey", [], {operator: "and"}
     assert_search "fresh honey", ["fresh", "honey"], {operator: :or}
+    assert_search "fresh honey", ["fresh", "honey"], {operator: :or, body_options: {track_total_hits: true}}
+    assert_search "fresh honey", [], {operator: :or, fields: [:name], match: :phrase, body_options: {track_total_hits: true}}
   end
 
   def test_operator_scoring