+++ test/notifications_test.rb
@@ -2,7 +2,7 @@ require_relative "test_helper"
 
 class NotificationsTest < Minitest::Test
   def test_search
-    Product.search_index.refresh
+    Product.searchkick_index.refresh
 
     notifications = capture_notifications do
       Product.search("product").to_a