+++ test/reindex_v2_job_test.rb
@@ -3,10 +3,10 @@ require_relative "test_helper"
 class ReindexV2JobTest < Minitest::Test
   def test_create
     product = Searchkick.callbacks(false) { Product.create!(name: "Boom") }
-    Product.search_index.refresh
+    Product.searchkick_index.refresh
     assert_search "*", []
     Searchkick::ReindexV2Job.perform_now("Product", product.id.to_s)
-    Product.search_index.refresh
+    Product.searchkick_index.refresh
     assert_search "*", ["Boom"]
   end
 
@@ -16,7 +16,7 @@ class ReindexV2JobTest < Minitest::Test
     assert_search "*", ["Boom"]
     Searchkick.callbacks(false) { product.destroy }
     Searchkick::ReindexV2Job.perform_now("Product", product.id.to_s)
-    Product.search_index.refresh
+    Product.searchkick_index.refresh
     assert_search "*", []
   end
 end